create separate css file for the Heart component, adjust url
