+++ frontend/src/components/Heart/Heart.js
@@ -1,10 +1,10 @@
 import React from "react";
 import { FaRegHeart, FaHeart } from "react-icons/fa";
+import './Heart.css'
 
 function Heart({ kojeSrce, isLike }) {
-  console.log(kojeSrce.id)
   return (
-    <div className="pet-like">
+    <div className={kojeSrce.isLike ? 'pet-like liked' : 'pet-like'}>
       {kojeSrce.isLike ? (
         <FaHeart onClick={() => isLike(kojeSrce.id)} />
       ) : (
